+++ src/main/java/roomescape/advice/GlobalExceptionHandler.java
@@ -5,7 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import roomescape.advice.dto.ErrorResponse;
-import roomescape.exception.AuthenticationException;
+import roomescape.auth.exception.AuthenticationException;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
