+++ src/main/java/roomescape/auth/exception/AuthenticationException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.auth.exception;
 
 public abstract class AuthenticationException extends RuntimeException {
     public AuthenticationException(String message) {
