+++ src/main/java/roomescape/auth/exception/MemberAuthenticationException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.auth.exception;
 
 public class MemberAuthenticationException extends AuthenticationException {
     private static final String ERROR_MESSAGE = "인증이 되지 않은 유저입니다.";
