+++ src/main/java/roomescape/config/handler/AdminAuthenticationInterceptor.java
@@ -6,12 +6,11 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
 import roomescape.auth.dto.LoggedInMember;
 import roomescape.auth.dto.RequestCookies;
+import roomescape.auth.exception.AdminAuthenticationException;
 import roomescape.auth.service.AuthService;
-import roomescape.exception.AdminAuthenticationException;
 
 @Component
 public class AdminAuthenticationInterceptor implements HandlerInterceptor {
-
     private final AuthService authService;
 
     public AdminAuthenticationInterceptor(AuthService authService) {
